+++ Restaurants.Application/Restaurants/Commands/DeleteRestaurant/DeleteRestaurantCommandHandler.cs
@@ -1,13 +1,16 @@
 ﻿using MediatR;
 using Microsoft.Extensions.Logging;
+using Restaurants.Domain.Constants;
 using Restaurants.Domain.Entities;
 using Restaurants.Domain.Exceptions;
+using Restaurants.Domain.Interfaces;
 using Restaurants.Domain.Repositories;
 
 namespace Restaurants.Application.Restaurants.Commands.DeleteRestaurant;
 
 public class DeleteRestaurantCommandHandler(ILogger<DeleteRestaurantCommand> logger, 
-                                            IRestaurantsRepository restaurantsRepository) : IRequestHandler<DeleteRestaurantCommand>
+                                            IRestaurantsRepository restaurantsRepository,
+                                            IRestaurantAuthorizationService restaurantAuthorizationService) : IRequestHandler<DeleteRestaurantCommand>
 {
     public async Task Handle(DeleteRestaurantCommand request, CancellationToken cancellationToken)
     {
@@ -17,6 +20,11 @@ public class DeleteRestaurantCommandHandler(ILogger<DeleteRestaurantCommand> log
         if (restaurant is null)
         {
             throw new NotFoundException(nameof(Restaurant), request.Id.ToString());
+        }
+
+        if (!restaurantAuthorizationService.Authorize(restaurant, ResourceOperation.Delete))
+        {
+            throw new ForbiddenException();
         }
         await restaurantsRepository.Delete(restaurant, cancellationToken);
     }
